Moves files that change together into modules based on feature.
