+++ client/src/__test__/actions/TaskActions.spec.js
@@ -1,6 +1,6 @@
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import * as actions from '../../actions/TaskActions'
+import * as actions from '../../app/task/TaskActions'
 import * as types from '../../constants/ActionTypes'
 
 const middlewares = [thunk]
