+++ client/src/containers/Analysis.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Activity from '../components/Activity'
+import Activity from '../app/activity/Activity'
 
 class Analysis extends React.Component {
   render() {
