+++ client/src/modules/activity/components/Activity.js
