+++ client/src/modules/efficiency/components/Efficiency.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { getElapsedTime, calculateFocus, initGraphData, updateGraphData } from '../helpers'
+import { getElapsedTime, calculateFocus, initGraphData, updateGraphData } from '../../../helpers'
 import PerformanceGraph from './PerformanceGraph'
 import moment from 'moment'
