+++ client/src/modules/rootReducer.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
-import timer from './TimerReducers'
-import task from './TaskReducers'
-import efficiency from './EfficiencyReducers'
-import activity from './ActivityReducers'
+import timer from './timer/reducers'
+import task from './task/reducers'
+import efficiency from './efficiency/reducers'
+import activity from './activity/reducers'
 
 export default combineReducers({
   routing: routerReducer,
