+++ client/src/modules/task/reducers.js
@@ -1,4 +1,4 @@
-import { SET_TASK, SET_ALERT } from '../constants/actionTypes'
+import { SET_TASK, SET_ALERT } from '../../constants/actionTypes'
 
 const initialState = {
   name: "",
