+++ client/src/modules/timer/saga.js
@@ -2,8 +2,8 @@ import { actionChannel, call, take, put, race } from 'redux-saga/effects'
 import {
   START_TIMER,
   STOP_TIMER
-} from '../constants/actionTypes'
-import { tick } from '../actions/TimerActions'
+} from '../../constants/actionTypes'
+import { tick } from './actions'
 
 // wait :: Number -> Promise
 const wait = ms => (
